requirements.txt
fastapi==0.135.3
numpy==1.26.4
openai==2.8.1
pandas==2.2.3
pdfplumber==0.11.9
PyMuPDF==1.27.2.2
pydantic==2.11.9
python-dotenv==1.1.1
requests==2.32.3
supabase==2.28.3
uvicorn==0.44.0
apify-client==2.5.0
